+++ NEWS.md
@@ -1,6 +1,7 @@
 
 # OpenRepGrid 0.1.13 
 
+  * setting a rating value outside the defined scale range now throws an error
   * 'ratings' to access and replace grid ratings added
   * 'elements' added to get and set element names replaces 'getElementNames' and 'eNames' which have become deprecated.
   * 'constructs', 'leftpoles', and 'rightpoles' added to get and set construct poles replace 'getConstructNames' and 'cNames' which have become  deprecated.